third_party/WebKit/LayoutTests/bluetooth/readValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="bluetooth-helpers.js"></script>
<script>
test(function(t) { assert_true(window.testRunner instanceof Object); t.done(); },
     'window.testRunner is required for the following tests.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('GenericAccessAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService('generic_access');
  }).then(function(service) {
    return service.getCharacteristic(generic_access.device_name.name);
  }).then(function(characteristic) {
    testRunner.setBluetoothMockDataSet('EmptyAdapter');
    return characteristic.readValue().then(function() {
      assert_unreached('Device went out of range, should fail.');
    }, function(e) {
      assert_equals(e.name, 'NetworkError');
      assert_equals(e.message, 'Bluetooth Device is no longer in range.');
    });
  });
}, 'Device goes out of range. Reject with NetworkError.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('GenericAccessAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['generic_access']}]})
    .then(device => device.connectGATT())
    .then(gattServer => gattServer.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('device_name'))
    .then(characteristic => {
      testRunner.setBluetoothMockDataSet('MissingServiceGenericAccessAdapter');
      return characteristic.readValue().then(() => {
        assert_unreached('Service was removed, should fail.');
      }, error => {
        assert_equals(error.name, 'InvalidStateError');
        assert_equals(error.message, 'GATT Service no longer exists.');
      });
    });
}, 'Services gets removed. Reject with InvalidStateError.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('GenericAccessAdapter');
  return requestDeviceWithKeyDown({filters: [{services: ['generic_access']}]})
    .then(device => device.connectGATT())
    .then(gattService => gattService.getPrimaryService('generic_access'))
    .then(service => service.getCharacteristic('device_name'))
    .then(characteristic => {
      testRunner.setBluetoothMockDataSet('MissingCharacteristicGenericAccessAdapter');
      return characteristic.readValue().then(() => {
        assert_unreached('Characteristic was removed, should fail.');
      }, error => {
        assert_equals(error.name, 'InvalidStateError');
        assert_equals(error.message, 'GATT Characteristic no longer exists.');
      });
    });
}, 'Characteristic gets removed. Reject with InvalidStateError.');

// The following tests make sure the Web Bluetooth implementation
// responds correctly to the different types of errors the
// underlying platform might return for GATT operations.

// Each implementation maps these characteristics to specific code paths
// that result in different errors thus increasing code coverage
// when testing. Therefore some of these characteristics might not be useful
// for all implementations.
[{
  testName: 'GATT Error: Unknown.',
  uuid: errorUUID(0xA1),
  errorType: 'NotSupportedError',
  message: 'GATT Error Unknown.'
}, {
  testName: 'GATT Error: Failed.',
  uuid: errorUUID(0xA2),
  errorType: 'NotSupportedError',
  message: 'GATT operation failed for unknown reason.'
}, {
  testName: 'GATT Error: In Progress.',
  uuid: errorUUID(0xA3),
  errorType: 'NetworkError',
  message: 'GATT operation already in progress.'
}, {
  testName: 'GATT Error: Invalid Length.',
  uuid: errorUUID(0xA4),
  errorType: 'InvalidModificationError',
  message: 'GATT Error: invalid attribute length.'
}, {
  testName: 'GATT Error: Not Permitted.',
  uuid: errorUUID(0xA5),
  errorType: 'NotSupportedError',
  message: 'GATT operation not permitted.'
}, {
  testName: 'GATT Error: Not Authorized.',
  uuid: errorUUID(0xA6),
  errorType: 'SecurityError',
  message: 'GATT operation not authorized.'
}, {
  testName: 'GATT Error: Not Paired.',
  uuid: errorUUID(0xA7),
  // TODO(ortuno): Change to InsufficientAuthenticationError or similiar
  // once https://github.com/WebBluetoothCG/web-bluetooth/issues/137 is
  // resolved.
  errorType: 'NetworkError',
  message: 'GATT Error: Not paired.'
}, {
  testName: 'GATT Error: Not Supported.',
  uuid: errorUUID(0xA8),
  errorType: 'NotSupportedError',
  message: 'GATT Error: Not supported.'
}].forEach(testSpec => {
  sequential_promise_test(() => {
    testRunner.setBluetoothMockDataSet('FailingGATTOperationsAdapter');
    return requestDeviceWithKeyDown({filters: [{services: [errorUUID(0xA0)]}]})
      .then(device => device.connectGATT())
      .then(gattServer => gattServer.getPrimaryService(errorUUID(0xA0)))
      .then(service => service.getCharacteristic(testSpec.uuid))
      .then(function(characteristic) {
        return characteristic.readValue().then(function() {
          assert_unreached('Trying to read the characteristic fails. Promise should reject.');
        }, function(e) {
          assert_equals(e.name, testSpec.errorType);
          assert_equals(e.message, testSpec.message);
        });
      });
  }, testSpec.testName);
});

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('GenericAccessAdapter');
  return requestDeviceWithKeyDown({
    filters: [{services: ['generic_access']}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService('generic_access');
  }).then(function(service) {
    return service.getCharacteristic(generic_access.device_name.name);
  }).then(function(characteristic) {
    return characteristic.readValue();
  }).then(function(value) {
    var decoder = new TextDecoder('utf-8');
    var value_str = decoder.decode(value);
    assert_equals(value_str, 'Generic Access Device');
  });
}, 'Request for characteristic. Should return right characteristic');
</script>
